refactor: Rename TogglableSettingCard to EditableSettingCard
